+++ include/element/GT9.hpp
@@ -32,14 +32,14 @@ class GT9 : public MeshElement{
     public:
     GT9(std::vector<long> u_pos, ElementShape s, ProjectData* data);
 
-    virtual std::vector<double> get_k() const override;
-    virtual MeshNode* get_stresses(size_t node, const std::vector<double>& u) const override;
-    virtual MeshNode* get_internal_loads(size_t node, const std::vector<double>& u) const override;
+    virtual std::vector<float> get_k() const override;
+    virtual MeshNode* get_stresses(size_t node, const std::vector<float>& u) const override;
+    virtual MeshNode* get_internal_loads(size_t node, const std::vector<float>& u) const override;
     virtual size_t get_gmsh_element_type() const override{ return 2;};
 
     private:
     Material const * const mat;
-    double t;
+    float t;
 };
 
 }
