+++ src/project_data.cpp
@@ -77,7 +77,7 @@ ProjectData::ProjectData(std::string project_file){
 #endif
         std::string absolute_path = project_file.substr(0, last_slash+1);
         absolute_path.append(geom_path);
-        double scale;
+        float scale;
         if(this->log_data(doc, "scale", TYPE_DOUBLE, false)){
             scale = doc["scale"].GetDouble();
         } else {
@@ -99,7 +99,7 @@ ProjectData::ProjectData(std::string project_file){
                 logger::log_assert(mat.HasMember(s.c_str()), logger::ERROR, "missing material property: {}", s);
                 logger::log_assert(mat[s.c_str()].IsArray() || mat[s.c_str()].IsDouble(), logger::ERROR, "material property {} must be either a number or an array of numbers", s);
             }
-            std::vector<std::vector<double>> values(5);
+            std::vector<std::vector<float>> values(5);
             for(size_t i = 0; i < properties.size(); ++i){
                 if(mat[properties[i].c_str()].IsArray()){
                     const auto& a = mat[properties[i].c_str()].GetArray();
@@ -126,10 +126,10 @@ ProjectData::ProjectData(std::string project_file){
                 this->log_data(mat, s, TYPE_DOUBLE, true);
             }
             this->log_data(mat, "plane_stress", TYPE_BOOL, true);
-            double E = mat["E"].GetDouble();
-            double nu = mat["nu"].GetDouble();
-            double Smax = mat["Smax"].GetDouble();
-            double Tmax = mat["Tmax"].GetDouble();
+            float E = mat["E"].GetDouble();
+            float nu = mat["nu"].GetDouble();
+            float Smax = mat["Smax"].GetDouble();
+            float Tmax = mat["Tmax"].GetDouble();
             bool plane_stress = mat["plane_stress"].GetBool();
             this->material.reset(new material::LinearElasticIsotropic(E*1e9, nu, Smax*1e6, Tmax*1e6, plane_stress));
         }
@@ -147,10 +147,10 @@ ProjectData::ProjectData(std::string project_file){
                     this->log_data(pathf, "step", TYPE_DOUBLE, true);
                     this->log_data(pathf, "max_turn_angle", TYPE_DOUBLE, true);
                     this->log_data(pathf, "turn_options", TYPE_INT, true);
-                    double step = pathf["step"].GetDouble();
-                    double angle = pathf["max_turn_angle"].GetDouble();
+                    float step = pathf["step"].GetDouble();
+                    float angle = pathf["max_turn_angle"].GetDouble();
                     int choices = pathf["turn_options"].GetInt();
-                    double restriction = 0;
+                    float restriction = 0;
                     if(this->log_data(pathf, "restriction_size", TYPE_DOUBLE, false)){
                         restriction = pathf["restriction_size"].GetDouble();
                     }
@@ -158,9 +158,9 @@ ProjectData::ProjectData(std::string project_file){
                 } else if(pathf["type"] == "visibility_graph"){
                     this->log_data(pathf, "step", TYPE_DOUBLE, true);
                     this->log_data(pathf, "max_turn_angle", TYPE_DOUBLE, true);
-                    double step = pathf["step"].GetDouble();
-                    double angle = pathf["max_turn_angle"].GetDouble();
-                    double restriction = 0;
+                    float step = pathf["step"].GetDouble();
+                    float angle = pathf["max_turn_angle"].GetDouble();
+                    float restriction = 0;
                     if(this->log_data(pathf, "restriction_size", TYPE_DOUBLE, false)){
                         restriction = pathf["restriction_size"].GetDouble();
                     }
